IN-111 fix: place the AddPhoto outside of the form, paths and types, remove ToastContainer
