+++ src/features/profile/settings/generalInfo/api/locations.api.ts
@@ -1,9 +1,10 @@
+import { ProfileEndpoints } from '@/features/profile/settings/generalInfo/api/profile.endpoints'
 import axios from 'axios'
 
 class LocationsApi {
   constructor() {}
   public async fetchCountries(): Promise<CountryData[]> {
-    return axios.get('/locations.json').then((res) => res.data)
+    return axios.get(ProfileEndpoints.locations).then((res) => res.data)
   }
 }
