+++ src/features/profile/settings/generalInfo/api/profile.endpoints.ts
@@ -1,0 +1,5 @@
+export const ProfileEndpoints = {
+  locations: '/locations.json',
+  profile: '/v1/users/profile',
+  uploadAvatar: '/v1/users/profile/avatar',
+}
