+++ src/features/profile/settings/generalInfo/model/generalInfoSchema.ts
@@ -1,11 +1,6 @@
 import { USER_NAME_REGEXP } from '@/features/auth'
 import { z } from 'zod'
 
-const imageSchema = z.object({
-  photo: z.string().nullable(),
-  photoForServer: z.instanceof(Blob).nullable(),
-})
-
 export const generalInfoSchema = z.object({
   aboutMe: z.string().max(200, 'Maximum length is 200 symbols').optional(),
   city: z.string().optional(),
@@ -19,7 +14,6 @@ export const generalInfoSchema = z.object({
     .string()
     .min(1, { message: 'Minimum length is 1 symbol' })
     .max(50, { message: 'Maximum length is 50 symbols' }),
-  photoData: imageSchema.optional(),
   userName: z
     .string()
     .min(6, { message: 'Minimum number of characters 6' })
