+++ src/features/profile/settings/generalInfo/model/profileStore.ts
@@ -4,11 +4,11 @@ import { createFileForUpload } from '@/common/utils/createFileForUpload'
 import { responseErrorHandler } from '@/common/utils/responseErrorHandler'
 import { PhotoResult } from '@/features/profile/settings/avatarDialog/model'
 import {
+  FormData,
   UpdatedProfile,
   UserProfile,
   profileAPi,
 } from '@/features/profile/settings/generalInfo/api'
-import { FormData } from '@/features/profile/settings/generalInfo/model/useFillGeneralForm'
 import { makeAutoObservable, runInAction } from 'mobx'
 
 class ProfileStore {
@@ -47,9 +47,6 @@ class ProfileStore {
       const res = await profileAPi.updateProfile(updatedData)
 
       if (res.status === 204) {
-        if (data.photoData) {
-          await this.uploadAvatar(data.photoData)
-        }
         toast.success('Your settings are saved!')
       }
     } catch (error) {
