+++ src/features/profile/settings/generalInfo/model/useFillGeneralForm.ts
@@ -1,29 +1,28 @@
+import { useState } from 'react'
 import { useForm } from 'react-hook-form'
 
 import { responseErrorHandler } from '@/common/utils/responseErrorHandler'
 import AuthStore from '@/features/auth/model/authStore'
 import { PhotoResult } from '@/features/profile/settings/avatarDialog/model'
+import { FormData } from '@/features/profile/settings/generalInfo'
 import { generalInfoSchema } from '@/features/profile/settings/generalInfo/model/generalInfoSchema'
 import ProfileStore from '@/features/profile/settings/generalInfo/model/profileStore'
 import { useFetchLocations } from '@/features/profile/settings/generalInfo/model/useFetchLocations'
 import { useHandleCalendar } from '@/features/profile/settings/generalInfo/model/useHandleCalendar'
 import { zodResolver } from '@hookform/resolvers/zod'
 
-export type FormData = {
-  aboutMe?: string
-  city?: string
-  country?: string
-  dateOfBirth?: string
-  firstName: string
-  lastName: string
-  photoData?: PhotoResult | undefined
-  userName: string
-}
-
 export const useFillGeneralForm = () => {
   const profile = AuthStore.profile
   const userProfile = ProfileStore.userProfile
-  const avatar = userProfile?.avatars[0].url
+  const currentPhoto = ProfileStore.userProfile?.avatars[0]?.url
+  const [photoObj, setPhotoObj] = useState<PhotoResult>({
+    photo: currentPhoto || null,
+    photoForServer: null,
+  })
+
+  const onModalPhotoSave = (photo: PhotoResult) => {
+    setPhotoObj(photo)
+  }
 
   const {
     control,
@@ -38,14 +37,6 @@ export const useFillGeneralForm = () => {
       dateOfBirth: userProfile?.dateOfBirth || '',
       firstName: userProfile?.firstName || '',
       lastName: userProfile?.lastName || '',
-      // photoData: {
-      //   photo: null,
-      //   photoForServer: null,
-      // },
-      photoData: avatar
-        ? { photo: avatar, photoForServer: null }
-        : { photo: null, photoForServer: null },
-
       userName: userProfile?.userName || profile?.userName,
     },
     mode: 'onChange',
@@ -59,6 +50,9 @@ export const useFillGeneralForm = () => {
   const onSubmit = async (data: FormData) => {
     try {
       await ProfileStore.updateProfile(data)
+      if (currentPhoto !== photoObj.photo) {
+        await ProfileStore.uploadAvatar(photoObj)
+      }
     } catch (error) {
       responseErrorHandler(error)
     }
@@ -74,7 +68,9 @@ export const useFillGeneralForm = () => {
     isCalendarOpen,
     isSubmitting,
     isValid,
+    onModalPhotoSave,
     onSubmit,
+    photoObj,
     selectDateHandler,
     toggleCalendar,
   }
