+++ src/features/profile/settings/generalInfo/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './AddPhoto'
+export * from './GeneralInfo'
